refactor: inventory controllers
