+++ package.json
@@ -19,6 +19,7 @@
       "express": "^4.17.1",
       "file-type": "^14.4.0",
       "graphql-request": "^1.8.2",
+      "http-status-codes": "^2.1.4",
       "logger": "^0.0.1",
       "moment": "^2.27.0",
       "moment-timezone": "^0.5.31",
